api/requirements.txt
flask>=1.0.2
numpy>=1.15.1
opencv-python>=4.0.0
pillow>=5.4.1
torch>=1.0.0
torchvision>=0.2.1